Add a biased variants of a textureLod() miplevel tests
These tests have the same purpose as the
{vs,fs}-textureLod-miplevels.shader_test ones. These are actually a
slightly modified copies of those. The modifications are setting the
lod_bias parameter to 1.0 and updating the expected colors in probes.

This exposes a bug in softpipe, where it does not take lod_bias into
account, so the colors returned in these tests are the same as in
unbiased versions.